+++ h2/src/main/org/h2/expression/SequenceValue.java
@@ -9,15 +9,13 @@ import org.h2.command.Prepared;
 import org.h2.engine.Session;
 import org.h2.message.DbException;
 import org.h2.schema.Sequence;
-import org.h2.table.ColumnResolver;
-import org.h2.table.TableFilter;
 import org.h2.value.TypeInfo;
 import org.h2.value.Value;
 
 /**
  * Wraps a sequence when used in a statement.
  */
-public class SequenceValue extends Expression {
+public class SequenceValue extends Operation0 {
 
     private final Sequence sequence;
 
@@ -62,29 +60,9 @@ public class SequenceValue extends Expression {
     }
 
     @Override
-    public void mapColumns(ColumnResolver resolver, int level, int state) {
-        // nothing to do
-    }
-
-    @Override
-    public Expression optimize(Session session) {
-        return this;
-    }
-
-    @Override
-    public void setEvaluatable(TableFilter tableFilter, boolean b) {
-        // nothing to do
-    }
-
-    @Override
     public StringBuilder getSQL(StringBuilder builder, int sqlFlags) {
         builder.append(current ? "CURRENT" : "NEXT").append(" VALUE FOR ");
         return sequence.getSQL(builder, sqlFlags);
-    }
-
-    @Override
-    public void updateAggregate(Session session, int stage) {
-        // nothing to do
     }
 
     @Override
